upgrade pelikan to the latest ccommon

+++ deps/ccommon/src/cc_print.c
@@ -82,7 +82,7 @@ _vscnprintf(char *buf, size_t size, const char *fmt, va_list args)
         return 0;
     }
 
-    if (i < size) {
+    if (i < (int)size) {
         return i;
     }
